+++ client/src/index.js
@@ -1,11 +1,14 @@
 import React, { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App";
+import App from "./components/App";
 import "tailwindcss/tailwind.css";
+import {BrowserRouter as Router} from 'react-router-dom'
 
 const root = createRoot(document.getElementById("root"));
 root.render(
-  <StrictMode>
-    <App />
-  </StrictMode>
+  <Router>
+    <StrictMode>
+      <App />
+    </StrictMode>
+  </Router>
 );
