+++ client/src/styles/LoginPage.css
@@ -1,5 +1,5 @@
 body {
-  background: url('bg.jpg');
+  background: url('../assets/images/background.jpg');
   background-size: cover;
   color: #083b13;
 }
